+++ src/product_logging/mod.rs
@@ -1,0 +1,4 @@
+//! Modules for product logging
+
+pub mod framework;
+pub mod spec;