+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Modules for log aggregation added ([#517]).
+
+[#517]: https://github.com/stackabletech/operator-rs/pull/517
+
 ## [0.28.8] - 2022-12-08
 
 ### Added